procedure/data file status

+++ TRIOS-MOBILE/FirstViewController.m
@@ -171,7 +171,7 @@
 {
 }
 
--(long)tableView:tableView numberOfRowsInSection:(NSInteger)section
+-(int)tableView:tableView numberOfRowsInSection:(NSInteger)section
 {
    return [status.names count];
 }
